feat: Extract error reporting methods to their own files
