+++ android/src/main/java/com/infomaniak/nativeonboarding/RNInfomaniakOnboardingView.kt
@@ -44,14 +44,11 @@ import expo.modules.kotlin.viewevent.EventDispatcher
 import expo.modules.kotlin.views.ExpoView
 import kotlinx.coroutines.launch
 
-private const val SUCCESS_EVENT_KEY = "accessTokens"
-private const val ERROR_EVENT_KEY = "error"
-
-class RNInfomaniakOnboardingView(context: Context, appContext: AppContext) : ExpoView(context, appContext) {
+class RNInfomaniakOnboardingView(context: Context, appContext: AppContext) : ExpoView(context, appContext), OnboardingEvents {
     // Creates and initializes an event dispatcher for the `onLoginSuccess` and `onLoginError` events.
     // The name of the event is inferred from the value and needs to match the event name defined in the module.
-    private val onLoginSuccess by EventDispatcher()
-    private val onLoginError by EventDispatcher()
+    override val onLoginSuccess by EventDispatcher()
+    override val onLoginError by EventDispatcher()
 
     private var loginData: LoginData? by mutableStateOf(null)
     private val pages = mutableStateListOf<Page>()
@@ -130,11 +127,11 @@ class RNInfomaniakOnboardingView(context: Context, appContext: AppContext) : Exp
         pages.addAll(config.slides)
 
         pages.forEach { page ->
-            page.backgroundImage.reportErrors()
+            page.backgroundImage.reportErrors(context)
 
             when (val illustration = page.animatedIllustration ?: page.staticIllustration) {
-                is AnimatedIllustration -> illustration.reportErrors()
-                is StaticIllustration -> illustration.reportErrors()
+                is AnimatedIllustration -> illustration.reportErrors(context)
+                is StaticIllustration -> illustration.reportErrors(context)
                 null -> reportMissingIllustration()
             }
         }
@@ -218,37 +215,6 @@ class RNInfomaniakOnboardingView(context: Context, appContext: AppContext) : Exp
 
     private fun stopLoadingLoginButtons() {
         areButtonsLoading = false
-    }
-
-    private fun reportAccessToken(accessToken: String) {
-        onLoginSuccess(mapOf(SUCCESS_EVENT_KEY to listOf(accessToken)))
-    }
-
-    private fun reportAccessTokens(accessTokens: List<String>) {
-        if (accessTokens.isNotEmpty()) {
-            onLoginSuccess(mapOf(SUCCESS_EVENT_KEY to accessTokens))
-        }
-    }
-
-    private fun reportError(errorMessage: String) {
-        onLoginError(mapOf(ERROR_EVENT_KEY to errorMessage))
-    }
-
-    private fun reportMissingAsset(fileName: String) {
-        reportError("Missing file '$fileName' in Android assets folders")
-    }
-
-    private fun reportMissingIllustration() {
-        reportError("Missing illustration. Provide either a StaticIllustration or an AnimatedIllustration instance in each slide")
-    }
-
-    private fun AnimatedIllustration.reportErrors() {
-        if (fileName.assetFileExists(context).not()) reportMissingAsset(fileName)
-    }
-
-    private fun StaticIllustration.reportErrors() {
-        if (androidDarkFileName.assetFileExists(context).not()) reportMissingAsset(androidDarkFileName)
-        if (androidLightFileName.assetFileExists(context).not()) reportMissingAsset(androidLightFileName)
     }
 
     companion object {
